ctrigf: correct variable types to float

+++ libm/complexf/internal/ctrigf.c
@@ -26,9 +26,9 @@ void __ccoshsinhf(float x, float *c, float *s)
 /* Program to subtract nearest integer multiple of PI */
 
 /* extended precision value of PI: */
-static const double DP1 =  3.140625;
-static const double DP2 =  9.67502593994140625E-4;
-static const double DP3 =  1.509957990978376432E-7;
+static const float DP1 =  3.140625;
+static const float DP2 =  9.67502593994140625E-4;
+static const float DP3 =  1.509957990978376432E-7;
 
 float __redupif(float x)
 {
